bpf: move map in map declarations to bpf.h
